+++ libs/core/algorithms/tests/unit/algorithms/mismatch_binary.cpp
@@ -5,154 +5,14 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/local/init.hpp>
-#include <hpx/modules/testing.hpp>
-#include <hpx/parallel/algorithms/mismatch.hpp>
 
-#include <cstddef>
 #include <iostream>
-#include <iterator>
-#include <numeric>
-#include <random>
 #include <string>
-#include <utility>
 #include <vector>
 
-#include "test_utils.hpp"
+#include "mismatch_binary_tests.hpp"
 
 ///////////////////////////////////////////////////////////////////////////////
-int seed = std::random_device{}();
-std::mt19937 gen(seed);
-std::uniform_int_distribution<> dis(0, 10006);
-
-template <typename IteratorTag>
-void test_mismatch_binary1(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    iterator begin1 = iterator(std::begin(c1));
-    iterator end1 = iterator(std::end(c1));
-
-    {
-        auto result = hpx::mismatch(begin1, end1, std::begin(c2), std::end(c2));
-
-        // verify values
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), c1.size());
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            c2.size());
-    }
-
-    {
-        std::size_t changed_idx = dis(gen);    //-V104
-        ++c1[changed_idx];
-
-        auto result = hpx::mismatch(begin1, end1, std::begin(c2), std::end(c2));
-
-        // verify values
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), changed_idx);
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            changed_idx);
-    }
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_mismatch_binary1(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    iterator begin1 = iterator(std::begin(c1));
-    iterator end1 = iterator(std::end(c1));
-
-    {
-        auto result =
-            hpx::mismatch(policy, begin1, end1, std::begin(c2), std::end(c2));
-
-        // verify values
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), c1.size());
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            c2.size());
-    }
-
-    {
-        std::size_t changed_idx = dis(gen);    //-V104
-        ++c1[changed_idx];
-
-        auto result =
-            hpx::mismatch(policy, begin1, end1, std::begin(c2), std::end(c2));
-
-        // verify values
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), changed_idx);
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            changed_idx);
-    }
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_mismatch_binary1_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    iterator begin1 = iterator(std::begin(c1));
-    iterator end1 = iterator(std::end(c1));
-
-    {
-        auto f = hpx::mismatch(p, begin1, end1, std::begin(c2), std::end(c2));
-        f.wait();
-
-        // verify values
-        auto result = f.get();
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), c1.size());
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            c2.size());
-    }
-
-    {
-        std::size_t changed_idx = dis(gen);    //-V104
-        ++c1[changed_idx];
-
-        auto f = hpx::mismatch(p, begin1, end1, std::begin(c2), std::end(c2));
-        f.wait();
-
-        // verify values
-        auto result = f.get();
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), changed_idx);
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            changed_idx);
-    }
-}
-
 template <typename IteratorTag>
 void test_mismatch_binary1()
 {
@@ -176,139 +36,6 @@ void mismatch_binary_test1()
 
 ///////////////////////////////////////////////////////////////////////////////
 template <typename IteratorTag>
-void test_mismatch_binary2(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    iterator begin1 = iterator(std::begin(c1));
-    iterator end1 = iterator(std::end(c1));
-
-    {
-        auto result = hpx::mismatch(begin1, end1, std::begin(c2), std::end(c2),
-            std::equal_to<std::size_t>());
-
-        // verify values
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), c1.size());
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            c2.size());
-    }
-
-    {
-        std::size_t changed_idx = dis(gen);    //-V104
-        ++c1[changed_idx];
-
-        auto result = hpx::mismatch(begin1, end1, std::begin(c2), std::end(c2),
-            std::equal_to<std::size_t>());
-
-        // verify values
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), changed_idx);
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            changed_idx);
-    }
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_mismatch_binary2(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    iterator begin1 = iterator(std::begin(c1));
-    iterator end1 = iterator(std::end(c1));
-
-    {
-        auto result = hpx::mismatch(policy, begin1, end1, std::begin(c2),
-            std::end(c2), std::equal_to<std::size_t>());
-
-        // verify values
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), c1.size());
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            c2.size());
-    }
-
-    {
-        std::size_t changed_idx = dis(gen);    //-V104
-        ++c1[changed_idx];
-
-        auto result = hpx::mismatch(policy, begin1, end1, std::begin(c2),
-            std::end(c2), std::equal_to<std::size_t>());
-
-        // verify values
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), changed_idx);
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            changed_idx);
-    }
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_mismatch_binary2_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    iterator begin1 = iterator(std::begin(c1));
-    iterator end1 = iterator(std::end(c1));
-
-    {
-        auto f = hpx::mismatch(p, begin1, end1, std::begin(c2), std::end(c2),
-            std::equal_to<std::size_t>());
-        f.wait();
-
-        // verify values
-        auto result = f.get();
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), c1.size());
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            c2.size());
-    }
-
-    {
-        std::size_t changed_idx = dis(gen);    //-V104
-        ++c1[changed_idx];
-
-        auto f = hpx::mismatch(p, begin1, end1, std::begin(c2), std::end(c2),
-            std::equal_to<std::size_t>());
-        f.wait();
-
-        // verify values
-        auto result = f.get();
-        HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), changed_idx);
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
-            changed_idx);
-    }
-}
-
-template <typename IteratorTag>
 void test_mismatch_binary2()
 {
     using namespace hpx::execution;
@@ -330,123 +57,6 @@ void mismatch_binary_test2()
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-template <typename IteratorTag>
-void test_mismatch_binary_exception(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::mismatch(iterator(std::begin(c1)), iterator(std::end(c1)),
-            std::begin(c2), std::end(c2), [](std::size_t, std::size_t) {
-                return throw std::runtime_error("test"), true;
-            });
-
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<hpx::execution::sequenced_policy,
-            IteratorTag>::call(hpx::execution::seq, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_mismatch_binary_exception(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::mismatch(policy, iterator(std::begin(c1)), iterator(std::end(c1)),
-            std::begin(c2), std::end(c2), [](std::size_t, std::size_t) {
-                return throw std::runtime_error("test"), true;
-            });
-
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(policy, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_mismatch_binary_exception_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    bool caught_exception = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        auto f =
-            hpx::mismatch(p, iterator(std::begin(c1)), iterator(std::end(c1)),
-                std::begin(c2), std::end(c2), [](std::size_t, std::size_t) {
-                    return throw std::runtime_error("test"), true;
-                });
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(p, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-    HPX_TEST(returned_from_algorithm);
-}
-
 template <typename IteratorTag>
 void test_mismatch_binary_exception()
 {
@@ -470,85 +80,7 @@ void mismatch_binary_exception_test()
     test_mismatch_binary_exception<std::forward_iterator_tag>();
 }
 
-/////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_mismatch_binary_bad_alloc(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    bool caught_bad_alloc = false;
-    try
-    {
-        hpx::mismatch(policy, iterator(std::begin(c1)), iterator(std::end(c1)),
-            std::begin(c2), std::end(c2), [](std::size_t, std::size_t) {
-                return throw std::bad_alloc(), true;
-            });
-
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_mismatch_binary_bad_alloc_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c1(10007);
-    std::vector<std::size_t> c2(c1.size());
-
-    std::size_t first_value = gen();    //-V101
-    std::iota(std::begin(c1), std::end(c1), first_value);
-    std::iota(std::begin(c2), std::end(c2), first_value);
-
-    bool caught_bad_alloc = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        auto f =
-            hpx::mismatch(p, iterator(std::begin(c1)), iterator(std::end(c1)),
-                std::begin(c2), std::end(c2), [](std::size_t, std::size_t) {
-                    return throw std::bad_alloc(), true;
-                });
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-    HPX_TEST(returned_from_algorithm);
-}
-
+///////////////////////////////////////////////////////////////////////////////
 template <typename IteratorTag>
 void test_mismatch_binary_bad_alloc()
 {
